+++ packages/yoga/src/Heading/Heading.theme.js
@@ -1,15 +1,17 @@
 const Heading = ({ spacing, colors }) => ({
   background: colors.white,
   padding: {
-    vertical: spacing.small,
-    horizontal: spacing.huge,
-    horizontalMobile: spacing.medium,
+    vertical: spacing.xxxsmall,
+    horizontal: spacing.small,
   },
   height: spacing.xxxlarge,
   button: {
-    width: spacing.large,
-    height: spacing.large,
-    marginRight: spacing.medium,
+    color: colors.secondary,
+    width: spacing.xxlarge + spacing.xxsmall,
+    height: spacing.xxlarge + spacing.xxsmall,
+    backgroundWidth: spacing.xxlarge,
+    backgroundHeight: spacing.xxlarge,
+    background: colors.elements.lineAndBorders,
   },
 });