+++ packages/yoga/src/NavigationMenu/web/BottomItems/BottomItems.tsx
@@ -3,6 +3,7 @@ import styled, { css } from 'styled-components';
 import { media } from '@gympass/yoga-helpers';
 
 const StyledItemsContainer = styled.nav`
+  position: relative;
   ${media.lg`display: none`}
 `;
 
@@ -11,22 +12,25 @@ const StyledItems = styled.ul`
     theme: {
       yoga: {
         components: {
-          navigationmenu: { backgroundColor, border, gap, padding },
+          navigationmenu: { backgroundColor, border, gap },
         },
       },
     },
   }) => css`
+    position: absolute;
+    bottom: 0;
     display: grid;
     grid-auto-columns: minmax(0, 1fr);
     grid-auto-flow: column;
     align-items: center;
     width: 100%;
-    height: 74px;
-    background-color: ${backgroundColor.default};
+    height: 56px;
+    background-color: ${backgroundColor.white};;
     margin: 0;
-    padding: ${padding.xxxsmall}px;
+    padding: 0;
     border: 1px solid ${border.color.default};
     gap: ${gap.xxxsmall}px;
+    z-index: 15;
   `}
 `;